Refactor common generator parts
