+++ WalletWasabi.Fluent.Generators/AutoNotifySyntaxReceiver.cs
@@ -1,0 +1,6 @@
+namespace WalletWasabi.Fluent.Generators;
+
+internal class AutoNotifySyntaxReceiver : FieldsWithAttributeSyntaxReceiver
+{
+	public override string AttributeClass => AutoNotifyGenerator.AutoNotifyAttributeDisplayString;
+}
