+++ WalletWasabi.Fluent.Generators/OaphSyntaxReceiver.cs
@@ -1,0 +1,6 @@
+namespace WalletWasabi.Fluent.Generators;
+
+internal class OaphSyntaxReceiver : FieldsWithAttributeSyntaxReceiver
+{
+	public override string AttributeClass => OapthGenerator.OaphAttributeDisplayString;
+}
